Reorganize into a simple bundle builder style
