+++ README.md
@@ -1,10 +1,25 @@
 # Good job, team!
+
 The congratulations we rightly deserve
 
 [gooodjob.team](https://gooodjob.team)
 
+## Running for Development
+
+```bash
+> yarn
+> yarn run start
+```
+
+## Building for Release
+
+```bash
+> yarn
+> yarn run build
+```
+
 ## Licenses
 
-* favicon.png - Licensed under Creative Commons with attribution, see details [here](https://commons.wikimedia.org/wiki/File:Emojione_1F389.svg#Licensing)
-* confetti.js - Copied from [here](https://codepen.io/jonathanbell/pen/OvYVYw), with a few slight modifications, licensed under MIT
-* All other files - Licensed under MIT, see LICENSE in this repository
+- favicon.png - Licensed under Creative Commons with attribution, see details [here](https://commons.wikimedia.org/wiki/File:Emojione_1F389.svg#Licensing)
+- confetti.js - Copied from [here](https://codepen.io/jonathanbell/pen/OvYVYw), with a few slight modifications, licensed under MIT
+- All other files - Licensed under MIT, see LICENSE in this repository
