+++ index.html
@@ -1,56 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-	<title>Good Job Team!</title>
-	<meta name="viewport" content="width=device-width initial-scale=1">
-	<meta name="description" content="The congratulations we rightly deserve">
-	<link rel="shortcut icon" href="favicon.png">
-	<style>
-		body {
-			background: #2a3340;
-			border: 0;
-			margin: 0;
-		}
-
-		#container {
-			display: flex;
-		}
-
-		#motivational-sentence {
-			font-size: 3em;
-			position: absolute;
-			top: 50%;
-			left: 50%;
-			transform: translate(-50%, -50%);
-			color: white;
-		}
-
-		#confetti-canvas {
-			overflow-y: hidden;
-			overflow-x: hidden;
-			width: 100%;
-			margin: 0;
-		}
-
-		#about {
-			position: absolute;
-            font-weight: bold;
-			bottom: 0;
-			right: 0;
-			color: white;
-			margin: 5px;
-		}
-	</style>
-	<script src="motivational_sentence.js" defer></script>
-	<script src="confetti.js" defer></script>
-</head>
-<body>
-<div id="container">
-	<h1 id="motivational-sentence">
-		<noscript>Good Job Team!</noscript>
-	</h1>
-	<canvas id="confetti-canvas"></canvas>
-</div>
-<a id="about" href="https://github.com/sburba/goodjobteam">about</a>
-</body>
+  <head>
+    <title>Good Job Team!</title>
+    <meta name="viewport" content="width=device-width initial-scale=1" />
+    <meta name="description" content="The congratulations we rightly deserve" />
+    <link rel="shortcut icon" href="static/favicon.png" />
+    <script src="src/index.js" defer></script>
+  </head>
+  <body>
+    <div id="container">
+      <h1 id="motivational-sentence">
+        <noscript>Good Job Team!</noscript>
+      </h1>
+      <canvas id="confetti-canvas"></canvas>
+    </div>
+    <a id="about" href="https://github.com/sburba/goodjobteam">about</a>
+  </body>
 </html>
