+++ src/index.js
@@ -1,0 +1,3 @@
+import "./styles.css";
+import "./confetti";
+import "./motivational_sentence";
