+++ src/styles.css
@@ -1,0 +1,34 @@
+body {
+  background: #2a3340;
+  border: 0;
+  margin: 0;
+}
+
+#container {
+  display: flex;
+}
+
+#motivational-sentence {
+  font-size: 3em;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  color: white;
+}
+
+#confetti-canvas {
+  overflow-y: hidden;
+  overflow-x: hidden;
+  width: 100%;
+  margin: 0;
+}
+
+#about {
+  position: absolute;
+  font-weight: bold;
+  bottom: 0;
+  right: 0;
+  color: white;
+  margin: 5px;
+}
